Pull test result embedder task implementation into its own gradle file
